tmux: Add keybinding to open small vertical to the right of the current one
I often need to do this because my monitor is very large, and without
the additional left padding in the form of that pane I'd have to turn my
head to look at code. Besides, the pane is very useful for running
various commands or programs like bacon for Rust.