Implement toggling dark/light mode in terminal and nvim
This is a very janky solution for easily toggling between light and dark
mode. If done via a Wezterm keybind, the Wezterm color scheme will
change, and the Neovim background will be set to either light or dark
upon next opening it. Running instances will not be affected.

Neovim just checks a file in $HOME/.cache after being open so it might
be slow? I don't really know, works so far, so I'm happy